Fixed and fully tested the feature for printing all history transactions
Account class
Added the property _currency and applied it the costructor and the method CreateNewAccount()
Added the property decimal Balance to set and get the decimal _balance. This one will be called after making any change on the value for balance.
Add a field ' List<string[]> _transaction '  to store the transactions.
Updated the method PrintAccountHistory() so that it prints content in _transaction.
Method for applying the changes for balances: SubmitTransaction(Customer customer, int accountIndex, decimal value)
Changed the string format for printing the balance to {_balance:0.00} , so that the Console will print '0' instead of blank if the value is 0

Bank class
Since the method for making a transfer wasn't ready, I had to make 3 methods to check if the transactions history work properly :  Deposit()   Withdraw()  VerifyCustomer()

Users class
For verifing a customer before making a transfer / withdrawla Added an overload method of CheckPassword() :   CheckPassword(Customer customer, string inputPassword)