Made some maintance in Customer, Account and StoreAndLoad classes.
Concluded the work with the text-files and their backups. The program will now load from the backup files in case a file is missing.   Worth to mention that I made a third constructor that works exclusively for declaring the objects after data loading.

Fully tested the entire program functionality and it's now prepared for implementing the Get request.