Auto currency exchange for transfering with updated exchange rates are now possible
Account class
thye currency list now holds 8 currencies encluding the price for swedish krona.

Admin class
Made the mehod SetCurrencyRate()  to able the admin to set the exchange price manually. And made the option for it in the Admin menu to choose the function.

Bank class
UpdateEchangeRates()  for making API-Calls to update the stored exchange rates.
GetRequest() to build the link request.
PrintCurrentExchange()  which it explains itself by the name.
ExchangeCurrency( transfer , currency) this one will convert the transfer from any currency to SEK for the sender account .
ExchangeBack(( transfer , currency) ) This will convert  to the recipient account.