The program can now perform external transactions with 15-min delay time.
Account class
New method MakeExternalTransfer() which works almost exactly like MakeTransfer, but it takes the part where the transfer is to be received and adds it as a task to  List<Task>.  I applied this method in ExternalTransfer() in the bank class.

Bank class
Made a timer and a list of tasks.  I set the timer for 900 000 milliseconds (15-min), before it loops and execute a task in the tasks list.