History transactions can now be loaded and saved.  User remove option for admins
Account class
SortOutDetails() method for sorting out the transaction for each account that being declared when the program loads.
Updated SubmitTransaction() so it can store a new transaction in a list that exports the changes to a text-file.

Admin class
DeleteUser() to enable admins to delete any user, (customer or admin).

StoreAndLoad class
LoadTransactions() to import form a text-file.
SaveTransactions() to export to a text-file.

I made some random history transactoins to have something to show in each account history